Extract scraped classes to own files
